Improve abstraction of Game type
I realized we can make Game a template for an abstract class, and ChessGame a
concrete derived class of the abstract class Game<ChessState, ChessMove> (and
also typedefs ChessState, ChessMove in its file). Still, the programmer has to
know to apply GameSolver<Game> only to a Game that is derived from an
application of Game<State, Move> (such as ChessGame), or it won't compile since
the Game will be missing the functions expected from a derived class of a
Game<State, Move>. But this is significantly cleaner because now Game<State,
Move> enforces the interface for its derived classes rather than the programmer
having to make sure that ChessGame, etc. fit a magical interface that is solely
specified by the functions GameSolver happens to use.

Now the Game template should specify the actual interface for a game, including
the enum return value of isOver() and the vector<Move> return value of
availableMoves(). The ChessGame implementation is of course still a stub.